requirements.txt
langchain
langchain_core
langchain_community
langchain_groq
python-dotenv
groq
mysql-connector-python
fastapi
uvicorn[standard]
ipykernel
streamlit
python-box
plotly